% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foxp2_sm.R
\docType{data}
\name{foxp2_sm}
\alias{foxp2_sm}
\title{Shortened Simulated FoxP2 Data Set.}
\format{
An object of class \code{data.frame} with 69 rows and 6 columns.
}
\usage{
foxp2_sm
}
\description{
A shortened version of the foxp2 data set for demonstrating R examples.
See details of the foxp2 data set by calling ?foxp2.
}
\keyword{datasets}
