% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_plot_results.R
\name{get_histogram_by_component}
\alias{get_histogram_by_component}
\title{Duration Times Plots: Histogram of duration times by Component}
\usage{
get_histogram_by_component(results, comp = 1:results$K)
}
\arguments{
\item{results}{results of duration times, i.e., results$results_duration}

\item{comp}{a vector of the components to plot; default is 1:isi_num_comp}
}
\value{
No return value, called for plotting histogram for each mixture component.
}
\description{
Plots the histogram of each mixture component superimposed the gamma distribution with shape and rate
parameters taken from the last MCMC iteration
}
\examples{
results <- BMRMM(foxp2_sm,num_cov=2,simsize=50)
get_histogram_by_component(results$results_duration)

}
